requirements.txt
discord-py-slash-command
discord-together
discord.py
python-dotenv
